compute-embeddings-function/requirements.txt
google-cloud-firestore==2.16.0
functions-framework==3.3.0
google-events==0.4.0
google-api-core
protobuf==4.22.1
cloudevents==1.9.0
google-cloud-aiplatform==1.44.0 